feat: Implement Command Blocks (#1223)
* Fix /transfer permission

* It's actually permission level three, not permission level two.

* Move dispatcher.register call to correct place

* Implement command blocks

* satisfy clippy

* satisfy formatter

* fix NBT typo

* Fix command block entity handling & ignore starting "/"

* fix: use @ for command block sender name

* feat: correctly load command block entities

* feat: send message when command block is changed

* feat: implement conditional flag for command blocks

* feat: implement chain command blocks

* fix: resolve clippy errors

* feat: implement command block previous output mode

* feat: implement comparator output for command blocks

* fix: check redstone conditions for repeating command blocks

---------

+++ pumpkin-util/src/translation.rs
@@ -168,12 +168,17 @@ pub fn get_translation_text<P: Into<Cow<'static, str>>>(
     }
 
     let (substitutions, indices) = reorder_substitutions(&translation, with);
+
+    // i32 since displacement can be negative. Casting from usize to i32 is safe in this context
+    // since it is unreasonable for the length of `range` or `substitution` to exceed 2147483647
+    // (i.e. i32::MAX).
     let mut displacement = 0i32;
     for (idx, &range) in indices.iter().enumerate() {
         let sub_idx = idx.clamp(0, substitutions.len() - 1);
         let substitution = substitutions[sub_idx].clone().get_text(locale);
         translation.replace_range(
-            range.start + displacement as usize..=range.end + displacement as usize,
+            (range.start as i32 + displacement) as usize
+                ..=(range.end as i32 + displacement) as usize,
             &substitution,
         );
         displacement += substitution.len() as i32 - range.len() as i32;
